[sway] add custom grimshot script
This is a modified version of the grimshot script that ships with
SwayWM (/usr/share/sway/scripts/grimshot).
It adds an 'upload' action. Calling this grimshot with 'upload' as the
action will behave the same as the 'save' (to file) action, but additionally
uploads to linode S3, using `linode-cli` (which must be preconfigured).