[sway] better window decoration colors
This now makes window border colors a littler more consistent
with Encom as well as more useful.

The focused, active window has the primary teal color for its border,
which makes it easy to know focus without being distracting,
(as was the case with the previous yellow).
The unfocused window is now a muted gray, which should make it more
obvious which window is in focus.
Lastly, the "indicator" color is purple. This changes the color of the
border edge that will be used for the next child window. e.g., when
tiling, a new window with vertical split enabled will open on the right
of the youngest child, which will now show a purple border on that
windows' right edge.