[sway] use `cliphist` for image support in clipboard
Updates sway keybinds and autostart to use
[Cliphist](https://github.com/sentriz/cliphist) instead of clipman.
Cliphist supports images, and for the most part seems to have parity
with clipman for simple line-based history viewing via fuzzel (or
anything dmenu-like). TBD if this works well long-term or with large
clipboard history, but it's promising.

FYI the cliphist database is store in ~/.cache/cliphist/db
and appears to use the very fast and light BoltDB pkg.