Create zenbook-duo.yml for 'way-displays'
Way-displays is an output profile config utility that serves as an alternative to `kanshi`. I found the configuration to be a little easier to use. This adds a config for that makes it easier to use the dual displays on a zenbook pro duo 14 OLED laptop. The main things are : setting the column layout so that sway understands this is a top and bottom display, setting hz to 120 for the oled and 60 for IPS bottom display.