[fish] allow easy window titles via variable
When using wezterm, and probably others, it's not easy to change the
terminal's window / tab title via OSC sequences, because of the default
behavior in fish_title.fish.

See https://github.com/wez/wezterm/issues/1593 for some explanation.

With this, we can easily override the title with an env var or
`set TERM_TITLE 'My New Title'`. The inpetus here was to differentiate
tabs (Editor, dev, files, etc.) when multiplex ssh'd into a dev server.