[helix] add whitespace color.
As pointed out on reddit, with whitespace rendering on, the default is
far too high contrast with the background, and matches the text too
closely making it distracting.

This adds a new shade, slightly darker than 'primary4', and sets the
indent guide to 'primary4' for a little pop on those guides.

This is, if anything, perhaps too low contrast with the background. Some
glyphs, such as carriage return (or probably newline, more accurately)
are maybe too difficult to identify quickly.

Color was picked using https://www.colorhexa.com/46595c