more host-specific configuration
Removes the need to symlink certain configuration, such as the
theme-specific waybar configuration by matching the hostname to the file
in sway/hosts/<hostname> . It's not perfect, but i think it's better
than needing to manually move or symlink files later.

This also updates more of the encom theme to the newer slate blue color.