[sway] attempt to improve touchpad feel.
IDK if something changed in recent kernel updates, or something
broader with endeavourOS, but the touchpad on my asus zenbook 13 has
felt ... off lately. Mostly, it has been slow to respond to initial
movements.

Using `wev` tool to measure pointer positioning didn't reveal anything
catastropihically wrong with the hardware.

The attempted improvement here is to turn off adaptive acceleration,
and adjust the acceleration ratio manually. Overall, this feels better
to me although I am now over-shooting some of the pointer targets, but
i'm hoping that is due to not being used to the new pointer tracking.