[sway] back to wezterm default
ghostty is still very inconsistent in rendering cursor. The cursor isn't
scaled correctly, and perhaps because of terminal line height, the
cursor glyph is cut off at the bottom.
I suspect this is largely a GTK issue related to HiDPI / scaling, but
nevertheless it's actually pretty annoying when trying to select text.

Also: the tab bar UI is yucky.

Also also: there's a glitchy skewing effect when animating the text
during a window size change. This happens often, e.g. when closing a
child window.