[waybar] fix shifting height
I was noticing that the waybar, and thus the sway windows / river views
would shift sometimes, when switching between workspaces (/ tags).

Using the debugging features in waybar, it seems to be caused by modules
changing their required height, likely due to excess padding-bottom.
The window module is the most likely culprit since it changes quite a
bit when switching workspaces.

The fix here is to reduce padding and declare a taller min-height of
28px to get the look close to what it was before.