[fish] update colors to look more encom
hmm, so fish_config has a GUI for setting colors, but all that does is
hardcode the color values as fish_variables. Welp, i guess i have to
keep comitting to that.

saving the fish_config in the GUI also seemed to update a bunch of other
things in the config.fish.

Also remove the color scheme from `btmb` alias, since that was
deprecated.